added some recomendations
