+++ app/middlewares/errors.js
@@ -8,7 +8,7 @@ const statusCodes = {
   [errors.DEFAULT_ERROR]: 500,
   [errors.SERVICE_ERROR]: 504,
   [errors.VALIDATE_ERROR]: 400,
-  [errors.COMPARE_ERROR]: 401
+  [errors.AUTHENTICATION_ERROR]: 401
 };
 
 exports.handle = (error, req, res, next) => {
