+++ config/constants/errorMessages.js
@@ -1,3 +1,3 @@
 exports.DB_CONNECTION = 'Connection failed';
 exports.GENERATE_TOKEN = 'Could not encode token';
-exports.COMPARE_PASS = 'Password does not match, try again.';
+exports.AUTHENTICATION = 'Password does not match, try again.';
